rename threshold to trap
